refactor: rename firestore.ts -> firestore-api.ts
